Performance profiles: default to 'classic' values
1st contact with gerrit, please tell me if stuff is wrong.

Changed the performance profiles so that default sets low limit 122MHz, 
which was standard before we had profiles. 
Seems to work well with 122, no need to raise that 
(and incur a battery penalty? If Arm is like Intel, CPU speed matters little _unless_ the voltage changes).

performance profile 0 sets upper limit to 768mhz. 
Again, to _me_ it seems good enough, when your prio is to 
keep the phone alive.

performance profile 2 stays at 360/1400.
360 may help people with audio dropouts, 1400 is a good stable overclock.

Change-Id: I51ab594db49e3c552ea5c4d297aa55c4c7bb373a