Replace cpu-hogging spinning animation in boot dialog with more modest one
When booting, new apps are optimized with dex2oat. During this process
a progress dialog is displayed (optimizing app 13/74). This dialog
has an indetemninate progress spinner that uses 50% of cpu.

This leaves only 50% for dex2oat, prolonging the time to boot.

With this lightweight bitmap replacement animation, dex2oat gets more
than 90% cpu.

This spinner is used in other places as well (e.g. shutdown)

Needs verification if it works on mdpi devices.

Change-Id: I5e25b45612a7bc61b61efb81680e76159f81c962